CathraAnimDemo
Imported Cathra's model and some of the textures. Imported all of the animations and made sure those work. There were a few that didn't (melee and ranged because of the lack of weapon models I think). Made an anim BP with entering combat, idle, run/walk, and jump. Pushing to ensure... stuff. Can't think of the word. Just say yes, we'll move on.