Made the pub, broke Cathra
The BaseCharacter is all messed up, but only for the TestLevel; works fine in everything else.

Making a video takes a lot of time, something my job doesn't really give me. So I made you a thing that shows what I expect (except the inside of the pub, which would be a separate area). Basically, don't go overboard, you're just using shapes, this is all getting scrapped when we make the final levels. We just need enough to go "Hey, this is what it will look like"