Added more models, cloud saved more UVs
Behold, stuff! Also, for tree 3 (since it would have taken me a long time to properly UV unwrap because of all the weirdness, I cheated a bunch. After doing a lightmap unwrap, I applied the grass texture to the leaves, then an altered version of the grass texture to the bark.)